Moving JVM json schema validation logic to JVM platform
