+++ amf-webapi/jvm/src/main/scala/amf/plugins/domain/webapi/unsafe/JsonSchemaValidatorBuilder.scala
@@ -1,0 +1,9 @@
+package amf.plugins.domain.webapi.unsafe
+
+import amf.plugins.document.webapi.validation.remote.{JvmJsonSchemaValidator, PlatformJsonSchemaValidator}
+
+object JsonSchemaValidatorBuilder {
+
+  def apply(): PlatformJsonSchemaValidator = JvmJsonSchemaValidator
+
+}
