+++ amf-webapi/shared/src/main/scala/amf/plugins/domain/webapi/unsafe/JsonSchemaSecrets.scala
@@ -1,0 +1,7 @@
+package amf.plugins.domain.webapi.unsafe
+
+import amf.plugins.document.webapi.validation.remote.PlatformJsonSchemaValidator
+
+trait JsonSchemaSecrets {
+  val jsonSchemaValidator: PlatformJsonSchemaValidator = JsonSchemaValidatorBuilder()
+}
